+++ lib/Borq.js
@@ -7,11 +7,13 @@ const FacebookUtils = require('./FacebookUtils.js');
 const Locales = require('./utils/Locales.js');
 const Config = require('../config/Config.js');
 const Translations = require('./Translations.js');
+const Conversations = require('./Conversations.js');
 
 module.exports = {
   aggregate: Aggregate,
   facebook: Facebook,
   facebookUtils: FacebookUtils,
+  conversations: Conversations,
   services: Services,
   http: HTTP,
   localeUtils: Locales,
