+++ lib/FacebookUtils.js
@@ -1,25 +1,4 @@
 /** @module facebookUtils */
-const Config = require('../config/Config.js');
-
-const utterances = {
-  yes: new RegExp(/^(yes*|yea|yup|yep|ya|sure|ok|y|yeah|yah|ya|sim)/i),
-  no: new RegExp(/^(no*|nah|nope|n|tidak|não)/i),
-  interrupts:
-  new RegExp(/^(restart*|get_started|stop|quit|switch_[a-z][a-z])/i),
-  greetings: ['help',
-              '👋',
-              'hello',
-              'start',
-              '(^restart(?!_))',
-              'mulai',
-              'halo',
-              'hi',
-              'oi',
-              'hai',
-              'membantu',
-              'socorro',
-              'olá'],
-};
 
 /**
 * Generates button objects
@@ -83,64 +62,9 @@ function generateQuickReply(text, replyArray) {
   };
 }
 
-/**
-* Handle breaks in the currently running conversation
-* @param {object} response the response object
-* @param {object} conversation the conversation object
-* @param {function} callback default function to call
-* @return {function} the function to direct the conversation
-*/
-function conversationSwitch(response, conversation, callback) {
-  if (utterances.interrupts.test(response.text)) {
-    return conversation.stop('interrupted');
-  } else {
-    return callback;
-  }
-}
-
-/**
-* Go to the next conversation.
-* @param {object} response the response object
-* @param {object} conversation the conversation object
-* @return {function} calling converationSwitch with conversation.next()
-*/
-function nextConversation(response, conversation) {
-  conversation.setTimeout(Config.conversationTimeout);
-  return conversationSwitch(response, conversation, conversation.next());
-}
-
-/**
-* Which conversation thread to go to
-* @param {string} threadName the name of the thread to go to next
-* @return {function} result of calling conversationSwitch with the threadName
-*/
-function goto(threadName) {
-  return (response, conversation) => {
-    conversation.setTimeout(Config.conversationTimeout);
-    conversationSwitch(response,
-                       conversation,
-                       conversation.gotoThread(threadName));
-  };
-}
-
-/**
-* Send a message to a facebookId, they have to have opted in
-* @param {object} bot a bot returned from calling `spawn`
-* @param {string} facebookId messenger id
-* @param {function} messageCallback callback to pass to startConversation
-*/
-function sendMessage(bot, facebookId, messageCallback) {
-  bot.startConversation({user: facebookId, channel: facebookId},
-                        messageCallback);
-}
-
 
 module.exports = {
   generateButtonObject,
-  nextConversation,
-  goto,
-  utterances,
   generateButtonTemplate,
   generateQuickReply,
-  sendMessage,
 };
